requirements.txt
google-generativeai>=0.8.0
flask>=3.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
werkzeug>=3.0.0
